requirements.txt
llama-index==0.11.4
llama-index-llms-ollama==0.3.1
llama-index-embeddings-ollama==0.3.0
python-dotenv==1.0.1